Revert "Add Data Hub Docmaps API v2 (#82)"
This reverts commit 13112c6dbd2be4412e17662af5839245fbdc2dbe.
