--- data_hub_api/docmaps/v2/codecs/preprint.py
@@ -1,41 +1,0 @@
-
-from data_hub_api.docmaps.v2.api_input_typing import ApiManuscriptVersionInput
-from data_hub_api.docmaps.v2.docmap_typing import (
-    DocmapAssertionItem,
-    DocmapPreprintInput
-)
-
-
-def get_docmap_preprint_input(
-    manuscript_version: ApiManuscriptVersionInput,
-    detailed: bool
-) -> DocmapPreprintInput:
-    if not detailed:
-        return {
-            'type': 'preprint',
-            'doi': manuscript_version['preprint_doi'],
-            'url': manuscript_version['preprint_url'],
-            'versionIdentifier': manuscript_version['preprint_version']
-        }
-    return {
-        'type': 'preprint',
-        'doi': manuscript_version['preprint_doi'],
-        'url': manuscript_version['preprint_url'],
-        'versionIdentifier': manuscript_version['preprint_version'],
-        'published': (
-            manuscript_version['preprint_published_at_date'].isoformat()
-            if manuscript_version['preprint_published_at_date']
-            else ''
-        ),
-        '_tdmPath': manuscript_version['tdm_path'],
-    }
-
-
-def get_docmap_preprint_assertion_item(
-    manuscript_version: ApiManuscriptVersionInput
-) -> DocmapAssertionItem:
-    return {
-        'type': 'preprint',
-        'doi': manuscript_version['preprint_doi'],
-        'versionIdentifier': manuscript_version['preprint_version']
-    }
