--- data_hub_api/docmaps/v2/sql/__init__.py
@@ -1,8 +1,0 @@
-import os.path
-
-
-SQL_DIR = os.path.dirname(__file__)
-
-
-def get_sql_path(sql_filename: str) -> str:
-    return os.path.join(SQL_DIR, sql_filename)
