+++ tests/unit_tests/docmaps/v1/codecs/elife_manuscript_test.py
@@ -1,12 +1,12 @@
-from data_hub_api.docmaps.v2.codecs.elife_manuscript import (
+from data_hub_api.docmaps.v1.codecs.elife_manuscript import (
     get_docmap_elife_manuscript_doi_assertion_item,
     get_docmap_elife_manuscript_output,
     get_elife_manuscript_version_doi
 )
 
-from tests.unit_tests.docmaps.v2.test_data import (
+from tests.unit_tests.docmaps.v1.test_data import (
     DOCMAPS_QUERY_RESULT_ITEM_1,
-    MANUSCRIPT_VERSION_1
+    PREPRINT_DETAILS_1
 )
 
 
@@ -29,15 +29,15 @@ class TestGetDocmapElifeManuscriptDoiAssertionItem:
     def test_should_populate_docmaps_elife_manuscript_doi_assertion_item(self):
         result = get_docmap_elife_manuscript_doi_assertion_item(
             query_result_item=DOCMAPS_QUERY_RESULT_ITEM_1,
-            manuscript_version=MANUSCRIPT_VERSION_1
+            preprint=PREPRINT_DETAILS_1
         )
         assert result == {
             'type': 'preprint',
             'doi': get_elife_manuscript_version_doi(
-                elife_doi_version_str=MANUSCRIPT_VERSION_1['elife_doi_version_str'],
+                elife_doi_version_str=PREPRINT_DETAILS_1['elife_doi_version_str'],
                 elife_doi=DOCMAPS_QUERY_RESULT_ITEM_1['elife_doi']
             ),
-            'versionIdentifier': MANUSCRIPT_VERSION_1['elife_doi_version_str']
+            'versionIdentifier': PREPRINT_DETAILS_1['elife_doi_version_str']
         }
 
 
@@ -45,16 +45,15 @@ class TestGetDocmapElifeManuscriptOutput:
     def test_should_populate_docmaps_elife_manuscript_output(self):
         result = get_docmap_elife_manuscript_output(
             query_result_item=DOCMAPS_QUERY_RESULT_ITEM_1,
-            manuscript_version=MANUSCRIPT_VERSION_1
+            preprint=PREPRINT_DETAILS_1
         )
         assert result == {
             'type': 'preprint',
             'identifier': DOCMAPS_QUERY_RESULT_ITEM_1['manuscript_id'],
             'doi': get_elife_manuscript_version_doi(
-                elife_doi_version_str=MANUSCRIPT_VERSION_1['elife_doi_version_str'],
+                elife_doi_version_str=PREPRINT_DETAILS_1['elife_doi_version_str'],
                 elife_doi=DOCMAPS_QUERY_RESULT_ITEM_1['elife_doi']
             ),
-            'versionIdentifier': MANUSCRIPT_VERSION_1['elife_doi_version_str'],
-            'published': '',
+            'versionIdentifier': PREPRINT_DETAILS_1['elife_doi_version_str'],
             'license': DOCMAPS_QUERY_RESULT_ITEM_1['license']
         }
