+++ tests/unit_tests/docmaps/v1/codecs/evaluation_test.py
@@ -21,7 +21,7 @@ from data_hub_api.docmaps.v1.codecs.evaluation import (
     get_elife_evaluation_doi_url,
     get_related_organization_detail
 )
-from tests.unit_tests.docmaps.v1.v1_test_data import (
+from tests.unit_tests.docmaps.v1.test_data import (
     ANNOTATION_CREATED_TIMESTAMP_1,
     DOCMAPS_QUERY_RESULT_ITEM_1,
     DOI_1,
