--- tests/unit_tests/docmaps/v1/codecs/v1_elife_manuscript_test.py
@@ -1,59 +1,0 @@
-from data_hub_api.docmaps.v1.codecs.elife_manuscript import (
-    get_docmap_elife_manuscript_doi_assertion_item,
-    get_docmap_elife_manuscript_output,
-    get_elife_manuscript_version_doi
-)
-
-from tests.unit_tests.docmaps.v1.v1_test_data import (
-    DOCMAPS_QUERY_RESULT_ITEM_1,
-    PREPRINT_DETAILS_1
-)
-
-
-class TestGetElifeVersionDoi:
-    def test_should_return_none_if_elife_doi_not_defined(self):
-        assert not get_elife_manuscript_version_doi(
-            elife_doi_version_str='elife_doi_version_str_1',
-            elife_doi=''
-        )
-
-    def test_should_return_elife_version_doi(self):
-        result = get_elife_manuscript_version_doi(
-            elife_doi_version_str='elife_doi_version_str_1',
-            elife_doi='elife_doi_1'
-        )
-        assert result == 'elife_doi_1.elife_doi_version_str_1'
-
-
-class TestGetDocmapElifeManuscriptDoiAssertionItem:
-    def test_should_populate_docmaps_elife_manuscript_doi_assertion_item(self):
-        result = get_docmap_elife_manuscript_doi_assertion_item(
-            query_result_item=DOCMAPS_QUERY_RESULT_ITEM_1,
-            preprint=PREPRINT_DETAILS_1
-        )
-        assert result == {
-            'type': 'preprint',
-            'doi': get_elife_manuscript_version_doi(
-                elife_doi_version_str=PREPRINT_DETAILS_1['elife_doi_version_str'],
-                elife_doi=DOCMAPS_QUERY_RESULT_ITEM_1['elife_doi']
-            ),
-            'versionIdentifier': PREPRINT_DETAILS_1['elife_doi_version_str']
-        }
-
-
-class TestGetDocmapElifeManuscriptOutput:
-    def test_should_populate_docmaps_elife_manuscript_output(self):
-        result = get_docmap_elife_manuscript_output(
-            query_result_item=DOCMAPS_QUERY_RESULT_ITEM_1,
-            preprint=PREPRINT_DETAILS_1
-        )
-        assert result == {
-            'type': 'preprint',
-            'identifier': DOCMAPS_QUERY_RESULT_ITEM_1['manuscript_id'],
-            'doi': get_elife_manuscript_version_doi(
-                elife_doi_version_str=PREPRINT_DETAILS_1['elife_doi_version_str'],
-                elife_doi=DOCMAPS_QUERY_RESULT_ITEM_1['elife_doi']
-            ),
-            'versionIdentifier': PREPRINT_DETAILS_1['elife_doi_version_str'],
-            'license': DOCMAPS_QUERY_RESULT_ITEM_1['license']
-        }
